requirements.txt
# ============================================================
# VoxEmotion – Requirements for Azure App Service
# Python 3.10.x
# ============================================================

# ── Web Server (REQUIRED for Azure — replaces Flask dev server) ─
gunicorn==21.2.0

# ── Deep Learning ────────────────────────────────────────────
torch==2.1.0
torchaudio==2.1.0

# ── Audio I/O ────────────────────────────────────────────────
soundfile==0.12.1
soxr==0.3.7
scipy==1.11.4
librosa

# ── Numerics ─────────────────────────────────────────────────
numpy==1.26.4

# ── Visualisation ────────────────────────────────────────────
matplotlib==3.8.2
seaborn==0.13.2

# ── Data / ML ─────────────────────────────────────────────────
pandas==2.1.4
scikit-learn==1.3.2
tqdm==4.66.1

# ── Web Framework ─────────────────────────────────────────────
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3

# ── File Parsing ──────────────────────────────────────────────
python-docx==1.1.2
PyPDF2==3.0.1
pdfminer.six==20231228

# ── Text Normalisation ────────────────────────────────────────
Unidecode==1.3.8
inflect==7.3.1

# ── Auth & Security ───────────────────────────────────────────
bcrypt==4.1.2
firebase-admin==6.5.0

# ── Azure Blob Storage (optional — for model file storage) ───
azure-storage-blob==12.19.0

# ── Misc ──────────────────────────────────────────────────────
setuptools>=69.0.0
python-dotenv
